+++ sdk/analytics/src/main/java/com/klaviyo/analytics/networking/requests/EventApiRequest.kt
@@ -2,7 +2,7 @@ package com.klaviyo.analytics.networking.requests
 
 import com.klaviyo.analytics.model.Event
 import com.klaviyo.analytics.model.Profile
-import com.klaviyo.coresdk.Registry
+import com.klaviyo.core.Registry
 import org.json.JSONObject
 
 /**
