+++ sdk/analytics/src/main/java/com/klaviyo/analytics/networking/requests/KlaviyoApiRequest.kt
@@ -1,6 +1,6 @@
 package com.klaviyo.analytics.networking.requests
 
-import com.klaviyo.coresdk.Registry
+import com.klaviyo.core.Registry
 import java.io.BufferedReader
 import java.io.IOException
 import java.io.InputStreamReader
