+++ sdk/core/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.klaviyo.coresdk">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.klaviyo.core">
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 </manifest>
