+++ sdk/core/src/main/java/com/klaviyo/core/config/Clock.kt
@@ -1,0 +1,15 @@
+package com.klaviyo.core.config
+
+interface Clock {
+
+    fun currentTimeMillis(): Long
+
+    fun currentTimeAsString(): String
+
+    fun schedule(delay: Long, task: () -> Unit): Cancellable
+
+    interface Cancellable {
+        fun runNow()
+        fun cancel(): Boolean
+    }
+}
