--- sdk/core/src/main/java/com/klaviyo/coresdk/Registry.kt
@@ -1,109 +1,0 @@
-package com.klaviyo.coresdk
-
-import com.klaviyo.coresdk.config.Clock
-import com.klaviyo.coresdk.config.Config
-import com.klaviyo.coresdk.config.KlaviyoConfig
-import com.klaviyo.coresdk.config.SystemClock
-import com.klaviyo.coresdk.lifecycle.KlaviyoLifecycleMonitor
-import com.klaviyo.coresdk.lifecycle.LifecycleMonitor
-import com.klaviyo.coresdk.model.DataStore
-import com.klaviyo.coresdk.model.SharedPreferencesDataStore
-import com.klaviyo.coresdk.networking.KlaviyoNetworkMonitor
-import com.klaviyo.coresdk.networking.NetworkMonitor
-import kotlin.reflect.KType
-import kotlin.reflect.typeOf
-
-class MissingDependency(type: KType) : Exception("No service registered for $type")
-class InvalidRegistration(type: KType) : Exception("Registered service does not match $type")
-
-typealias Registration = () -> Any
-
-/**
- * Services registry for decoupling SDK components
- * Acts as a very basic Service Locator for internal SDK dependencies
- *
- * Core dependencies are defined as properties for ease of access.
- * Dynamic dependencies can be defined and fetched with the add/get methods.
- * At this time I treat all our services as singletons
- *
- * Technical note on the dynamic registry methods:
- * - noinline keyword allows us to store the Registration lambdas
- * - inline methods make it possible to use generics properly here
- * - @PublishedApi is required so that the inlined methods can read/write to private registry
- *   without leaving the type-erased methods fully public
- */
-object Registry {
-
-    val configBuilder: Config.Builder get() = KlaviyoConfig.Builder()
-
-    val config: Config get() = get()
-
-    val clock: Clock = SystemClock
-
-    val lifecycleMonitor: LifecycleMonitor get() = KlaviyoLifecycleMonitor
-
-    val networkMonitor: NetworkMonitor get() = KlaviyoNetworkMonitor
-
-    val dataStore: DataStore get() = SharedPreferencesDataStore
-
-    /**
-     * Internal registry of registered service instances
-     */
-    @PublishedApi
-    internal val services = mutableMapOf<KType, Any>()
-
-    /**
-     * Internal registry of registered service lambdas
-     */
-    @PublishedApi
-    internal val registry = mutableMapOf<KType, Registration>()
-
-    /**
-     * Register a service for a type, specified by generic parameter
-     * Typical usage would be to register the singleton implementation of an interface
-     *
-     * @param T - Type, usually an interface, to register under
-     * @param service - The implementation
-     */
-    inline fun <reified T : Any> add(service: Any) {
-        val type = typeOf<T>()
-        services[type] = service
-        registry.remove(type)
-    }
-
-    /**
-     * Lazily register a service builder for a type, specified by generic parameter
-     * Typical usage would be to register a builder method for the implementation of an interface
-     *
-     * @param T - Type, usually an interface, to register under
-     * @param registration - Lambda that returns the implementation
-     */
-    inline fun <reified T : Any> add(noinline registration: Registration) {
-        val type = typeOf<T>()
-        services.remove(type)
-        registry[type] = registration
-    }
-
-    /**
-     * Get a registered service by type
-     *
-     * @param T - Type of service, usually an interface
-     * @return The instance of the service
-     * @throws MissingDependency If no service is registered of that type
-     */
-    inline fun <reified T : Any> get(): T {
-        val type = typeOf<T>()
-        val service: Any? = services[type]
-
-        if (service is T) return service
-
-        when (val s = registry[type]?.let { it() }) {
-            is T -> {
-                services[type] = s
-                return s
-            }
-            is Any -> throw InvalidRegistration(type)
-            else -> throw MissingDependency(type)
-        }
-    }
-}
