--- sdk/core/src/main/java/com/klaviyo/coresdk/config/Clock.kt
@@ -1,15 +1,0 @@
-package com.klaviyo.coresdk.config
-
-interface Clock {
-
-    fun currentTimeMillis(): Long
-
-    fun currentTimeAsString(): String
-
-    fun schedule(delay: Long, task: () -> Unit): Cancellable
-
-    interface Cancellable {
-        fun runNow()
-        fun cancel(): Boolean
-    }
-}
