+++ sdk/push-fcm/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.klaviyo.push_fcm" />
