+++ settings.gradle
@@ -1,3 +1,3 @@
 // sdk modules
 include ':sdk'
-include ':sdk:core', ':sdk:analytics', ':sdk:push', ':sdk:core-shared-tests'
+include ':sdk:core', ':sdk:analytics', ':sdk:push-fcm', ':sdk:core-shared-tests'
